Reactoring packages path
