+++ sdgx/manager.py
@@ -10,7 +10,7 @@ import pluggy
 from sdgx import models
 from sdgx.exceptions import InitializationError, NotFoundError, RegisterError
 from sdgx.log import logger
-from sdgx.utils.utils import Singleton
+from sdgx.utils import Singleton
 
 
 class Manager(metaclass=Singleton):
