+++ sdgx/models/components/sample.py
@@ -1,14 +1,12 @@
 """
-# Refer CTGAN Version 0.6.0: https://github.com/sdv-dev/CTGAN@a40570e321cb46d798a823f350e1010a0270d804
-# Which is Lincensed by MIT License
+Refer CTGAN Version 0.6.0: https://github.com/sdv-dev/CTGAN@a40570e321cb46d798a823f350e1010a0270d804
+Which is Lincensed by MIT License
 """
 
 import numpy as np
 
-from sdgx.data_processors.samplers.base import Sampler
 
-
-class DataSamplerCTGAN(Sampler):
+class DataSamplerCTGAN:
     """DataSampler samples the conditional vector and corresponding data for CTGAN."""
 
     def __init__(self, data, output_info, log_frequency):
