+++ sdgx/models/components/transform.py
@@ -11,8 +11,6 @@ import pandas as pd
 from joblib import Parallel, delayed
 from rdt.transformers import ClusterBasedNormalizer, OneHotEncoder
 
-from sdgx.data_processors.transformers.base import Transformer
-
 SpanInfo = namedtuple("SpanInfo", ["dim", "activation_fn"])
 ColumnTransformInfo = namedtuple(
     "ColumnTransformInfo",
@@ -20,7 +18,7 @@ ColumnTransformInfo = namedtuple(
 )
 
 
-class DataTransformer(Transformer):
+class DataTransformer:
     """Data Transformer.
 
     Model continuous columns with a BayesianGMM and normalized to a scalar [0, 1] and a vector.
