+++ sdgx/models/components/utils.py
@@ -1,47 +1,8 @@
 import contextlib
 import logging
-import threading
 
 import numpy as np
 import torch
-
-
-# 添加增加的日志log
-def get_log_file_handler(log_path):
-    # 具体参数请参考：
-    # https://docs.python.org/3/library/logging.handlers.html#timedrotatingfilehandler
-    # 查路径请参考：
-    # fileshandle.baseFilename
-    fileshandler = logging.handlers.TimedRotatingFileHandler(
-        log_path, when="W6", interval=5, backupCount=15, encoding="utf-8"
-    )
-    fileshandler.suffix = "%Y%m%d_%H%M%S.log"
-    fileshandler.setLevel(logging.DEBUG)
-    fmt_str = "%(asctime)s %(levelname)s %(filename)s[%(lineno)d] %(message)s"
-    formatter = logging.Formatter(fmt_str)
-    fileshandler.setFormatter(formatter)
-    return fileshandler
-
-
-# 引入自 ctgan
-# 未来会根据业务需求优化修改
-def random_state(function):
-    """Set the random state before calling the function.
-
-    Args:
-        function (Callable):
-            The function to wrap around.
-    """
-
-    def wrapper(self, *args, **kwargs):
-        if self.random_states is None:
-            return function(self, *args, **kwargs)
-
-        else:
-            with set_random_states(self.random_states, self.set_random_state):
-                return function(self, *args, **kwargs)
-
-    return wrapper
 
 
 @contextlib.contextmanager
@@ -73,6 +34,42 @@ def set_random_states(random_state, set_model_random_state):
 
         np.random.set_state(original_np_state)
         torch.set_rng_state(original_torch_state)
+
+
+def random_state(function):
+    """Set the random state before calling the function.
+
+    Args:
+        function (Callable):
+            The function to wrap around.
+    """
+
+    def wrapper(self, *args, **kwargs):
+        if self.random_states is None:
+            return function(self, *args, **kwargs)
+
+        else:
+            with set_random_states(self.random_states, self.set_random_state):
+                return function(self, *args, **kwargs)
+
+    return wrapper
+
+
+# 添加增加的日志log
+def get_log_file_handler(log_path):
+    # 具体参数请参考：
+    # https://docs.python.org/3/library/logging.handlers.html#timedrotatingfilehandler
+    # 查路径请参考：
+    # fileshandle.baseFilename
+    fileshandler = logging.handlers.TimedRotatingFileHandler(
+        log_path, when="W6", interval=5, backupCount=15, encoding="utf-8"
+    )
+    fileshandler.suffix = "%Y%m%d_%H%M%S.log"
+    fileshandler.setLevel(logging.DEBUG)
+    fmt_str = "%(asctime)s %(levelname)s %(filename)s[%(lineno)d] %(message)s"
+    formatter = logging.Formatter(fmt_str)
+    fileshandler.setFormatter(formatter)
+    return fileshandler
 
 
 def flatten_array(nested, prefix=""):
@@ -242,15 +239,3 @@ def validate_numerical_distributions(numerical_distributions, metadata_columns):
         #         f'{invalid_columns}. The column names you provide must be present '
         #         'in the metadata.'
         #     )
-
-
-class Singleton(type):
-    _instances = {}
-    _lock = threading.Lock()
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            with cls._lock:
-                if cls not in cls._instances:
-                    cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
