--- sdgx/utils/__init__.py
@@ -1,6 +1,0 @@
-import logging.config
-
-# from config import logger_config_path
-
-# logging.config.fileConfig(logger_config_path)
-logger = logging.getLogger("SDG")
